Enhanced _document with dynamic language support
This commit refines the _document file by incorporating dynamic language support. The currentLanguage is now retrieved from the Redux store using the useSelectorState hook from '@/src/core/hooks'. This improvement allows for the seamless adjustment of the 'lang' attribute in the HTML tag based on the current language set in the application.