fix: update lang attribute in _document.tsx to resolve React-Redux context error
This commit addresses a React-Redux context error that occurred due to the absence of the Redux Provider wrapping in the `_document.tsx` file. The error message, "could not find react-redux context value," pointed to a lack of access to the React-Redux context during the server-side rendering phase.

To resolve this issue, the `lang` attribute in `_document.tsx` was updated to a static value (`'es,en'`). This change eliminates the reliance on the `currentLanguage` state from the Redux Store within the `_document.tsx` file, ensuring a smoother server-side rendering process during the Next.js build. The error was a result of missing Redux context in this specific file, and this fix mitigates the problem by providing a direct language specification in the `lang` attribute.